Refactor test step definitions to allow for generic/specific test steps
